Дополнение №1
-Добавлены комментарии в контроллеры и некоторые модели
-Добавлены Feature тесты
1. Тесты на корректную регистрацию и авторизацию
2. Тесты на AdminMiddleware (для обычных пользователей и гостей)

+++ app/Http/Controllers/LikeController.php
@@ -9,7 +9,7 @@ class LikeController extends Controller
     public function store(Post $post)
     {
         $post->likes()->create([
-            'user_id' => auth()->id(),
+            'user_id' => auth()->id(),  //Создание лайка, используем auth()->id() для добавление его в базу под именем 'user_id'
         ]);
 
         return back();
@@ -17,7 +17,7 @@ class LikeController extends Controller
 
     public function destroy(Post $post)
     {
-        $post->likes()->where('user_id', auth()->id())->delete();
+        $post->likes()->where('user_id', auth()->id())->delete(); //Удаление лайка где user_id равен auth()->id()
         return back();
     }
 }
